requirements.txt
Django==1.9.13
django-cms==3.4
django-filer
djangocms_text_ckeditor
djangocms-link
djangocms-file
djangocms-picture
djangocms-video
djangocms-googlemap
djangocms-snippet
djangocms-style
djangocms-column
psycopg2==2.6.2
requests==2.13.0
tableauserverclient==0.4.1
